Move tests to existing test modules
