+++ tests/unit/test_config.py
@@ -560,6 +560,15 @@ class ConfigTestCase(TestCase, AdaptedConfigurationTestCaseMixin):
             os.path.join(tempdir, 'c')
         ]))
 
+    def test_validate_bad_file_roots(self):
+        expected = salt.config._expand_glob_path(
+            [salt.syspaths.BASE_FILE_ROOTS_DIR]
+        )
+        with mock.patch('salt.config._normalize_roots') as mk:
+            ret = salt.config._validate_file_roots(None)
+            assert not mk.called
+        assert ret == {'base': expected}
+
     @with_tempfile()
     @with_tempdir()
     def test_master_pillar_roots_glob(self, tempdir, fpath):
@@ -582,6 +591,15 @@ class ConfigTestCase(TestCase, AdaptedConfigurationTestCaseMixin):
             os.path.join(tempdir, 'b'),
             os.path.join(tempdir, 'c')
         ]))
+
+    def test_validate_bad_pillar_roots(self):
+        expected = salt.config._expand_glob_path(
+            [salt.syspaths.BASE_PILLAR_ROOTS_DIR]
+        )
+        with mock.patch('salt.config._normalize_roots') as mk:
+            ret = salt.config._validate_pillar_roots(None)
+            assert not mk.called
+        assert ret == {'base': expected}
 
     @with_tempdir()
     def test_master_id_function(self, tempdir):
